requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
cffi==1.15.1
channels==3.0.5
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.1
django-environ==0.9.0
django-rest-framework==0.1.0
django-simple-history==3.1.1
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
drf-extensions==0.7.1
entrypoints==0.4
executing==0.10.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipykernel==6.15.2
ipython==8.4.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg==3.1
psycopg-binary==3.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyodbc==4.0.34
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.2.1
pyzmq==23.2.1
reportlab==3.6.11
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
text-unidecode==1.3
tornado==6.2
traitlets==5.3.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
wcwidth==0.2.5
zope.interface==5.4.0
